Add BusinessExceptionHandler for Deleting Department

+++ web-springboot/tlias-study-management/src/main/java/com/albert/service/impl/DeptServiceImpl.java
@@ -1,6 +1,8 @@
 package com.albert.service.impl;
 
+import com.albert.exception.BusinessException;
 import com.albert.mapper.DeptMapper;
+import com.albert.mapper.EmpMapper;
 import com.albert.pojo.Dept;
 import com.albert.service.DeptService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -13,6 +15,8 @@ public class DeptServiceImpl implements DeptService {
 
     @Autowired
     private DeptMapper deptMapper;
+    @Autowired
+    private EmpMapper empMapper;
 
     @Override
     public List<Dept> findAll() {
@@ -21,6 +25,10 @@ public class DeptServiceImpl implements DeptService {
 
     @Override
     public void deleteById(Integer id) {
+        Integer count = empMapper.countByDeptId(id);
+        if(count > 0){
+            throw new BusinessException("部门下有员工， 不能删除");
+        }
         deptMapper.deleteById(id);
     }
